feat: configurable assets directory
this commit makes assets directory (--assets its equivalent SERVER_ASSETS) configurable.
it means that assets directory is no more restricted to be relative to root directory.
so relative and absolutes *paths* work.

because of the assets directory is treated as independent path.
it's necessary to adjust the assets directory path to be explicit.

for example: having the root `./my-root`
the assets directory path ´./my-assets´ should be changed to `./my-root/my-assets`