feat: worker threads multiplier option --threads-multiplier
A new -t, --threads-multiplier <threads-multiplier> option which
represent a number of worker threads multiplier that'll be multiplied
by the number of system CPUs using the formula:
worker threads = number of CPUs * n
where `n` is the value that changes with this option.
When the multiplier value is 0 or 1 then the `number of CPUs` is only used.
The number of worker threads result should be a number between 1 and
32,768 though it is advised to keep this value on the smaller side.