feat: auto-config file detection at startup
The TOML config file is now detected at startup time using a new
default `./config.toml` value. It makes sure to check this file path
and load it if found. Otherwise, the behavior continues as usual.

So it's not mandatory to provide a flag or env variable explicitly if
the default file path is wanted.

Also, additional debug information and file validations are provided.