refactor: drop linux ppc64le/s390x from alpine and scratch docker
The SWS binaries for ppc64le (powerpc64le-unknown-linux-gnu) and
s590x (s390x-unknown-linux-gnu) are dynamically linked so
it does **not** make so much sense to deliver them with either the
Alpine (musl) or the Scratch images (they even do not work properly as
of writing).

So we remove those two from the SWS Alpine and Scratch Docker image
variants to avoid misunderstandings or future issues.

The Debian Docker image should be preferred instead, either linux/ppc64le
or linux/s390x respectively.

This resolves #308