refactor: prevent single wildcards from matching path separators for URL Redirects (#501)
* Enable literal_separator for redirects

This ensures that * does not match the path separator.

* Fix tests

* Update documentation

BREAKING CHANGE: Up to version 2.33.1 the wildcard `*` was matching the path separator. For example, `/{*}/{*}/` matched `/assets/images/logo/`. In later versions, the default has changed such that `*` does not match the path separator. In contrast, double wildcard (`**`) can match also a path separator.