virtual hosts: support HTTP/2 using :authority
So far, the HTTP host header has been used to determine the hostname
that is being accessed. In HTTP/2, this header is not guaranteed to be
sent. Instead, the ":authority" pseudo-header field is used. This commit
adds a check for that header to make virtual hosts also work with
HTTP/2.

Additionally, this commit adds code to check and remove the port number,
which may or may not be sent as part of the HTTP host header.